turned BN integration test into API BN integration test
